fix: move lighting tools to their own module to avoid circular import
